:sparkles: feat: useColoredText, useValidation 훅
